os/board/rtl8721cms: Separate RTK mpu code structure with common code
